Improved quote demo and documentation, ready to go! 🍾
